requirements.txt
Flask==1.1.2
numpy==1.16.4
pyyaml==5.1.2
requests==2.22.0
pillow==6.1.0
matplotlib==3.1.1
torchvision==0.4.0
gunicorn
gevent